requirements.txt
streamlit
anthropic
chromadb
pymupdf
langchain
langchain-text-splitters
sentence-transformers
python-dotenv